requirements.txt
# These requirements are for DEVELOPMENT ONLY!
# You do not need e.g. Sphinx or Fudge just to run the 'fab' tool.
# Instead, these are necessary for executing the test suite.
#
# PLEASE NOTE: due to how Pip currently works, you will need to manually
# execute "pip install Mercurial" before using this file. Otherwise it will
# bail out upon trying to install Fudge from BitBucket.

pycrypto >= 1.9
nose
coverage
pylint
-e hg+http://bitbucket.org/kumar303/fudge/#egg=Fudge
Sphinx >= 0.6
